FIX(BUILD): Fix error of format-extra-args and format mismatch.
1. There is no format of org_name in output string.
2. _reclaim_region_count does not match format of %ul.

--story=866624711

Contributed-by:emoryzheng

    modified:   hotspot/src/share/vm/gc_implementation/g1/g1CollectedHeap.hpp
    modified:   hotspot/src/share/vm/memory/filemap.cpp

+++ hotspot/src/share/vm/gc_implementation/g1/g1CollectedHeap.hpp
@@ -237,7 +237,7 @@ private:
   // It keeps track of the humongous regions.
   HeapRegionSet _humongous_set;
 
-  int _reclaim_region_count;
+  size_t _reclaim_region_count;
 
   void eagerly_reclaim_humongous_regions();
